requirements.txt
cryptography==3.4.7
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.20.1
opencv-python==4.5.1.48
psycopg2-binary==2.8.6
python-dotenv==0.17.1